hyperHTML without adoption capabilities.
hyper.adopt(...) has been experimental for a while
and it never worked properly in all target browsers.

Instead of bloating hyperHTML core
adopt should be somehow handled externally
so that only when used where it makes sense,
with viperHTML compatible template literals,
it makes sense to include adopt functionality too.

This version of hyperHTML is back to 3.9k instead of > 4k.